+++ main/shared/src/CAmbiguity.cpp
@@ -5,7 +5,7 @@
 // *******************************************************************************************
 
 #include <iostream>
-#include <String.h>
+#include <string.h>
 #include "CDataFile.h"
 #include "CAmbiguity.h"
 #include "CRoseCode.h"
@@ -17,35 +17,35 @@
 // The root constructor is called by the respective inherited constructor.
 CAmbRoot::CAmbRoot(char* fileName)
 {
-	int			fileLength;
-    CRoseCode	infoBlock;
-	CDataFile	ambFile;
-	
-	ambFile.OpenPath(DATA_FILE_PATH, fileName);
-	fileLength = ambFile.GetLength();
-	ambTblPtr = (CRoseCode*) ambFile.ReadAll();
-	ambFile.Close();
+    long        fileLength;
+    CRoseCode   infoBlock;
+    CDataFile   ambFile;
+    
+    ambFile.OpenPath(DATA_FILE_PATH, fileName);
+    fileLength = ambFile.GetLength();
+    ambTblPtr = (CRoseCode*) ambFile.ReadAll();
+    ambFile.Close();
 #ifndef __BIG_ENDIAN__
-    int n = fileLength/sizeof(CRoseCode);
+    long n = fileLength/sizeof(CRoseCode);
     for (int i=0; i < n; i++) {
         ambTblPtr[i].EndianSwap();
     }
 #endif
 
-	// the information block at the end of the file says the dimensions of the 2D table
-	memcpy(&infoBlock, ambTblPtr + fileLength/sizeof(CRoseCode) - 1, sizeof(CRoseCode));
-	arraySize 		= infoBlock.GetArraySize();
-	maxReplacements = infoBlock.GetMaxReplacements();
+    // the information block at the end of the file says the dimensions of the 2D table
+    memcpy(&infoBlock, ambTblPtr + fileLength/sizeof(CRoseCode) - 1, sizeof(CRoseCode));
+    arraySize       = infoBlock.GetArraySize();
+    maxReplacements = infoBlock.GetMaxReplacements();
 
-	if (arraySize*maxReplacements*sizeof(CRoseCode) != fileLength-sizeof(CRoseCode))
-		throw CFatalError("CAmbRoot::CAmbRoot", "Ambiguity File Data Invalid"); 
+    if (arraySize*maxReplacements*sizeof(CRoseCode) != fileLength-sizeof(CRoseCode))
+        throw CFatalError("CAmbRoot::CAmbRoot", "Ambiguity File Data Invalid"); 
 }
 
 // ###########################################################################################
 CAmbRoot::~CAmbRoot()
 {
-	if (ambTblPtr != NULL)
-		delete[] (char*)ambTblPtr;
+    if (ambTblPtr != NULL)
+        delete[] (char*)ambTblPtr;
 }
 
 // ###########################################################################################
@@ -67,41 +67,41 @@ CIntAmbTable::CIntAmbTable() : CAmbRoot(INT_AMB_FILE)
 }
 
 // ###########################################################################################
-int	CExtAmbTable::FirstReplacement(CRoseCode& findMe)
+int CExtAmbTable::FirstReplacement(CRoseCode& findMe)
 {
-	searchCode = findMe;
-	iLine = searchCode.StripToExtBase();
-	iCol = 0;
-	findMe = searchCode | LookUp(iLine, iCol);
-	return 1;
+    searchCode = findMe;
+    iLine = searchCode.StripToExtBase();
+    iCol = 0;
+    findMe = searchCode | LookUp(iLine, iCol);
+    return 1;
 }
 
-int	CMidAmbTable::FirstReplacement(CRoseCode& findMe)
+int CMidAmbTable::FirstReplacement(CRoseCode& findMe)
 {
-	searchCode = findMe;
-	iLine = searchCode.StripToMidBase();
-	iCol = 0;
-	findMe = searchCode | LookUp(iLine, iCol);
-	return 1;
+    searchCode = findMe;
+    iLine = searchCode.StripToMidBase();
+    iCol = 0;
+    findMe = searchCode | LookUp(iLine, iCol);
+    return 1;
 }
 
-int	CIntAmbTable::FirstReplacement(CRoseCode& findMe)
+int CIntAmbTable::FirstReplacement(CRoseCode& findMe)
 {
-	searchCode = findMe;
-	iLine = searchCode.StripToIntBase();
-	iCol = 0;
-	findMe = searchCode | LookUp(iLine, iCol);
-	return 1;
+    searchCode = findMe;
+    iLine = searchCode.StripToIntBase();
+    iCol = 0;
+    findMe = searchCode | LookUp(iLine, iCol);
+    return 1;
 }
 
 // ###########################################################################################
-int	CAmbRoot::NextReplacement(CRoseCode& fixMe)
+int CAmbRoot::NextReplacement(CRoseCode& fixMe)
 {
-	if ((LookUp(iLine, ++iCol).BLANK_ROSE_CODE()) || (iCol >= maxReplacements))
-		return 0;
-		
-	fixMe = searchCode | LookUp(iLine, iCol);
-	return 1;
+    if ((LookUp(iLine, ++iCol).BLANK_ROSE_CODE()) || (iCol >= maxReplacements))
+        return 0;
+        
+    fixMe = searchCode | LookUp(iLine, iCol);
+    return 1;
 }
 
 // ###########################################################################################
